+++ src/tkn.ts
@@ -74,7 +74,7 @@ export interface Tkn {
   getRawClusterTasks(): Promise<TknTask[]>;
   execute(command: CliCommand, cwd?: string, fail?: boolean): Promise<CliExitData>;
   executeWatch(command: CliCommand, opts?: {}): WatchProcess;
-  executeInTerminal(command: CliCommand, cwd?: string): void;
+  executeInTerminal(command: CliCommand, resourceName?: string, cwd?: string): void;
   getTaskRunsForTasks(task: TektonNode): Promise<TektonNode[]>;
   getTaskRunsForClusterTasks(task: TektonNode): Promise<TektonNode[]>;
   getTriggerTemplates(triggerTemplates?: TektonNode): Promise<TektonNode[]>;
@@ -557,22 +557,33 @@ export class TknImpl implements Tkn {
     return data;
   }
 
-  async executeInTerminal(command: CliCommand, cwd: string = process.cwd(), name = 'Tekton'): Promise<void> {
-    let toolLocation = await ToolsConfig.detectOrDownload();
+  async executeInTerminal(command: CliCommand, resourceName?: string, cwd: string = process.cwd(), name = 'Tekton'): Promise<void> {
+    let toolLocation = await ToolsConfig.detectOrDownload(command.cliCommand);
     if (toolLocation) {
       toolLocation = path.dirname(toolLocation);
     }
-    const terminal: Terminal = WindowUtil.createTerminal(name, cwd, toolLocation);
+    let terminal: Terminal;
+    if (resourceName) {
+      terminal = WindowUtil.createTerminal(`${name}:${resourceName}`, cwd, toolLocation);
+    } else {
+      terminal = WindowUtil.createTerminal(name, cwd, toolLocation);
+    }
     terminal.sendText(cliCommandToString(command), true);
     terminal.show();
   }
 
   async execute(command: CliCommand, cwd?: string, fail = true): Promise<CliExitData> {
     if (command.cliCommand.indexOf('tkn') >= 0) {
-      const toolLocation = ToolsConfig.getTknLocation();
+      const toolLocation = ToolsConfig.getTknLocation('tkn');
       if (toolLocation) {
         // eslint-disable-next-line require-atomic-updates
         command.cliCommand = command.cliCommand.replace('tkn', `"${toolLocation}"`).replace(new RegExp('&& tkn', 'g'), `&& "${toolLocation}"`);
+      }
+    } else {
+      const toolLocation = await ToolsConfig.detectOrDownload(command.cliCommand);
+      if (toolLocation) {
+        // eslint-disable-next-line require-atomic-updates
+        command.cliCommand = command.cliCommand.replace(command.cliCommand, `"${toolLocation}"`).replace(new RegExp(`&& ${command.cliCommand}`, 'g'), `&& "${toolLocation}"`);
       }
     }
 
@@ -582,12 +593,10 @@ export class TknImpl implements Tkn {
   }
 
   executeWatch(command: CliCommand, cwd?: string): WatchProcess {
-    if (command.cliCommand.indexOf('tkn') >= 0) {
-      const toolLocation = ToolsConfig.getTknLocation();
-      if (toolLocation) {
-        // eslint-disable-next-line require-atomic-updates
-        command.cliCommand = command.cliCommand.replace('tkn', `"${toolLocation}"`).replace(new RegExp('&& tkn', 'g'), `&& "${toolLocation}"`);
-      }
+    const toolLocation = ToolsConfig.getTknLocation(command.cliCommand);
+    if (toolLocation) {
+      // eslint-disable-next-line require-atomic-updates
+      command.cliCommand = command.cliCommand.replace(command.cliCommand, `"${toolLocation}"`).replace(new RegExp(`&& ${command.cliCommand}`, 'g'), `&& "${toolLocation}"`);
     }
 
     return cli.executeWatch(command, cwd ? { cwd } : {});