+++ test/tekton/taskrun.test.ts
@@ -6,7 +6,6 @@
 'use strict';
 
 import * as os from 'os';
-import * as path from 'path';
 import * as fs from 'fs-extra';
 import * as vscode from 'vscode';
 import * as chai from 'chai';
@@ -16,9 +15,8 @@ import { TestItem } from './testTektonitem';
 import { TaskRun } from '../../src/tekton/taskrun';
 import { TknImpl } from '../../src/tkn';
 import { TektonItem } from '../../src/tekton/tektonitem';
-import { cli, CliExitData } from '../../src/cli';
+import { CliExitData } from '../../src/cli';
 import * as logInEditor from '../../src/util/log-in-editor';
-import { Platform } from '../../src/util/platform';
 import { ContextType } from '../../src/context-type';
 import { Command } from '../../src/cli-command';
 
@@ -28,7 +26,6 @@ chai.use(sinonChai);
 suite('Tekton/TaskRun', () => {
   const sandbox = sinon.createSandbox();
   let execStub: sinon.SinonStub;
-  let quote: string;
   let showErrorMessageStub: sinon.SinonStub;
   let unlinkStub: sinon.SinonStub;
   let getPipelineRunNamesStub: sinon.SinonStub;
@@ -38,7 +35,6 @@ suite('Tekton/TaskRun', () => {
   const taskRunItem = new TestItem(pipelineRunItem, 'taskrun', ContextType.PIPELINERUN, undefined, '2019-07-25T12:03:00Z', 'True');
   let showLogInEditorStub: sinon.SinonStub;
   let configurationStub: sinon.SinonStub;
-  let cliExecuteStub: sinon.SinonStub;
   let osStub: sinon.SinonStub;
   let writeFileStub: sinon.SinonStub;
   let showInformationMessageStub: sinon.SinonStub;
@@ -50,12 +46,10 @@ suite('Tekton/TaskRun', () => {
     sandbox.stub(TknImpl.prototype, 'getPipelineRunsList').resolves([pipelineRunItem]);
     sandbox.stub(TknImpl.prototype, 'getTasks').resolves([taskRunItem]);
     sandbox.stub(vscode.window, 'showInputBox').resolves();
-    cliExecuteStub = sandbox.stub(cli, 'execute');
     osStub = sandbox.stub(os, 'tmpdir').resolves();
     unlinkStub = sandbox.stub(fs, 'unlink').resolves();
     showInformationMessageStub = sandbox.stub(vscode.window, 'showInformationMessage').resolves();
     showErrorMessageStub = sandbox.stub(vscode.window, 'showErrorMessage').resolves();
-    quote = Platform.OS === 'win32' ? '"' : '\'';
     writeFileStub = sandbox.stub(fs, 'writeFile').resolves();
     showLogInEditorStub = sandbox.stub(logInEditor, 'showLogInEditor').resolves();
     configurationStub = sandbox.stub(vscode.workspace, 'getConfiguration').returns({ get: () => false } as unknown as vscode.WorkspaceConfiguration);
@@ -254,14 +248,13 @@ suite('Tekton/TaskRun', () => {
         }
       })});
       osStub.returns('path');
-      cliExecuteStub.resolves({ stdout: 'successfully created'});
+      execStub.onSecondCall().resolves({ stdout: 'successfully created'});
       await TaskRun.restartTaskRun(taskRunItem);
       unlinkStub.calledOnce;
       osStub.calledOnce;
       writeFileStub.calledOnce;
       showInformationMessageStub.calledOnce;
-      const mockFilePath = path.join('path', `${taskRunItem.getName()}-.yaml`);
-      expect(cliExecuteStub).calledOnceWith(Command.create(`${quote}${mockFilePath}${quote}`));
+      execStub.calledTwice;
       expect(showInformationMessageStub).calledOnceWith('TaskRun successfully restarted');
     });
 
@@ -296,14 +289,13 @@ suite('Tekton/TaskRun', () => {
         }
       })});
       osStub.returns('path');
-      cliExecuteStub.resolves({ error: 'fails'});
+      execStub.onSecondCall().resolves({ error: 'fails'});
       await TaskRun.restartTaskRun(taskRunItem);
       unlinkStub.calledOnce;
       osStub.calledOnce;
       writeFileStub.calledOnce;
       showErrorMessageStub.calledOnce;
-      const mockFilePath = path.join('path', `${taskRunItem.getName()}-.yaml`);
-      expect(cliExecuteStub).calledOnceWith(Command.create(`${quote}${mockFilePath}${quote}`));
+      execStub.calledTwice;
       expect(showErrorMessageStub).calledOnceWith('Fail to restart TaskRun: fails');
     });